restore libwasmvm to ease reference

+++ internal/api/mocks.go
@@ -19,11 +19,13 @@ import (
 
 const MOCK_CONTRACT_ADDR = "contract"
 
+// MockEnv returns a mock environment for testing
+// this is the original, and should not be changed.
 func MockEnv() types.Env {
 	return types.Env{
 		Block: types.BlockInfo{
 			Height:  123,
-			Time:    types.Uint64(1578939743_987654321),
+			Time:    1578939743_987654321,
 			ChainID: "foobar",
 		},
 		Transaction: &types.TransactionInfo{
